refactor: rename file and variables
