+++ lib/utils/DevServerPlugin.js
@@ -131,8 +131,8 @@ class DevServerPlugin {
       return defaultValue;
     };
 
-    const config = compiler.options;
-    config.plugins = config.plugins || [];
+    const compilerOptions = compiler.options;
+    compilerOptions.plugins = compilerOptions.plugins || [];
     /** @type {Boolean} */
     const webTarget = [
       'web',
@@ -142,17 +142,17 @@ class DevServerPlugin {
       // eslint-disable-next-line no-undefined
       undefined,
       null,
-    ].includes(config.target);
+    ].includes(compilerOptions.target);
     /** @type {Entry} */
     const additionalEntries = checkInject(
       options.injectClient,
-      config,
+      compilerOptions,
       webTarget
     )
       ? [clientEntry]
       : [];
 
-    if (hotEntry && checkInject(options.injectHot, config, true)) {
+    if (hotEntry && checkInject(options.injectHot, compilerOptions, true)) {
       additionalEntries.push(hotEntry);
     }
 
@@ -177,8 +177,14 @@ class DevServerPlugin {
         )
       );
     } else {
-      config.entry = prependEntry(config.entry || './src', additionalEntries);
-      compiler.hooks.entryOption.call(config.context, config.entry);
+      compilerOptions.entry = prependEntry(
+        compilerOptions.entry || './src',
+        additionalEntries
+      );
+      compiler.hooks.entryOption.call(
+        compilerOptions.context,
+        compilerOptions.entry
+      );
     }
 
     const providePlugin = new webpack.ProvidePlugin({
@@ -187,14 +193,14 @@ class DevServerPlugin {
     providePlugin.apply(compiler);
 
     if (
-      !config.plugins.find(
+      !compilerOptions.plugins.find(
         (p) => p.constructor === webpack.HotModuleReplacementPlugin
       )
     ) {
       // add and apply the HMR plugin, if it didn't exist before.
       if (options.hot === true || options.hot === 'only') {
         const plugin = new webpack.HotModuleReplacementPlugin();
-        config.plugins.push(plugin);
+        compilerOptions.plugins.push(plugin);
         plugin.apply(compiler);
       }
     }
