+++ lib/utils/updateCompiler.js
@@ -1,15 +1,15 @@
 'use strict';
 
-const DevServerPlugin = require('./addEntries');
+const DevServerPlugin = require('./DevServerPlugin');
 
 function updateCompiler(compiler, options) {
   const compilers = compiler.compilers || [compiler];
   // eslint-disable-next-line no-shadow
   compilers.forEach((compiler) => {
-    const config = compiler.options;
-    config.plugins = config.plugins || [];
+    const compilerOptions = compiler.options;
+    compilerOptions.plugins = compilerOptions.plugins || [];
 
-    let devServerPlugin = config.plugins.find(
+    let devServerPlugin = compilerOptions.plugins.find(
       (p) => p.constructor === DevServerPlugin
     );
     if (devServerPlugin) {
@@ -17,7 +17,7 @@ function updateCompiler(compiler, options) {
       devServerPlugin.options = options;
     } else {
       devServerPlugin = new DevServerPlugin(options);
-      config.plugins.push(devServerPlugin);
+      compilerOptions.plugins.push(devServerPlugin);
     }
     devServerPlugin.apply(compiler);
   });
