+++ test/server/Server.test.js
@@ -4,7 +4,7 @@ const { relative, sep } = require('path');
 const webpack = require('webpack');
 const sockjs = require('sockjs/lib/transport');
 const Server = require('../../lib/Server');
-const DevServerPlugin = require('../../lib/utils/addEntries');
+const DevServerPlugin = require('../../lib/utils/DevServerPlugin');
 const config = require('../fixtures/simple-config/webpack.config');
 const port = require('../ports-map').Server;
 const isWebpack5 = require('../helpers/isWebpack5');
